+++ src/lib/Timeline.js
@@ -68,6 +68,8 @@ export default class ReactCalendarTimeline extends Component {
 
     itemTouchSendsClick: PropTypes.bool,
 
+    horizontalLineClassNamesForGroup: PropTypes.func,
+
     onItemMove: PropTypes.func,
     onItemResize: PropTypes.func,
     onItemClick: PropTypes.func,
@@ -189,6 +191,8 @@ export default class ReactCalendarTimeline extends Component {
 
     traditionalZoom: false,
     showCursorLine: false,
+
+    horizontalLineClassNamesForGroup: null,
 
     onItemMove: null,
     onItemResize: null,
@@ -832,15 +836,17 @@ export default class ReactCalendarTimeline extends Component {
     this.props.onCanvasDoubleClick(groupId, time, e)
   }
 
-  horizontalLines(canvasWidth, groupHeights) {
+  horizontalLines(canvasWidth, groupHeights, groups) {
     return (
       <GroupRows
+        groups={groups}
         canvasWidth={canvasWidth}
         lineCount={_length(this.props.groups)}
         groupHeights={groupHeights}
         clickTolerance={this.props.clickTolerance}
         onRowClick={this.handleRowClick}
         onRowDoubleClick={this.handleRowDoubleClick}
+        horizontalLineClassNamesForGroup={this.props.horizontalLineClassNamesForGroup}
       />
     )
   }
@@ -1089,7 +1095,8 @@ export default class ReactCalendarTimeline extends Component {
     const { height, groupHeights, groupTops } = stackingMethod(
       dimensionItems,
       groupOrders,
-      lineHeight
+      lineHeight,
+      groups
     )
 
     return { dimensionItems, height, groupHeights, groupTops }
@@ -1299,7 +1306,7 @@ export default class ReactCalendarTimeline extends Component {
                 height,
                 headerHeight
               )}
-              {this.horizontalLines(canvasWidth, groupHeights)}
+              {this.horizontalLines(canvasWidth, groupHeights, groups)}
               {this.todayLine(
                 canvasTimeStart,
                 canvasTimeEnd,