+++ src/lib/row/GroupRow.js
@@ -8,7 +8,9 @@ class GroupRow extends Component {
     onDoubleClick: PropTypes.func.isRequired,
     isEvenRow: PropTypes.bool.isRequired,
     style: PropTypes.object.isRequired,
-    clickTolerance: PropTypes.number.isRequired
+    clickTolerance: PropTypes.number.isRequired,
+    group: PropTypes.object.isRequired,
+    horizontalLineClassNamesForGroup: PropTypes.func
   }
 
   render() {
@@ -17,14 +19,21 @@ class GroupRow extends Component {
       isEvenRow,
       style,
       onClick,
-      clickTolerance
+      clickTolerance,
+      horizontalLineClassNamesForGroup,
+      group
     } = this.props
+
+    let classNamesForGroup = [];
+    if (horizontalLineClassNamesForGroup) {
+      classNamesForGroup = horizontalLineClassNamesForGroup(group);
+    }
 
     return (
       <PreventClickOnDrag clickTolerance={clickTolerance} onClick={onClick}>
         <div
           onDoubleClick={onDoubleClick}
-          className={isEvenRow ? 'rct-hl-even' : 'rct-hl-odd'}
+          className={(isEvenRow ? 'rct-hl-even ' : 'rct-hl-odd ') + (classNamesForGroup ? classNamesForGroup.join(' ') : '')}
           style={style}
         />
       </PreventClickOnDrag>