+++ src/lib/row/GroupRows.js
@@ -9,14 +9,17 @@ export default class GroupRows extends Component {
     groupHeights: PropTypes.array.isRequired,
     onRowClick: PropTypes.func.isRequired,
     onRowDoubleClick: PropTypes.func.isRequired,
-    clickTolerance: PropTypes.number.isRequired
+    clickTolerance: PropTypes.number.isRequired,
+    groups: PropTypes.array.isRequired,
+    horizontalLineClassNamesForGroup: PropTypes.func
   }
 
   shouldComponentUpdate(nextProps) {
     return !(
       nextProps.canvasWidth === this.props.canvasWidth &&
       nextProps.lineCount === this.props.lineCount &&
-      nextProps.groupHeights === this.props.groupHeights
+      nextProps.groupHeights === this.props.groupHeights &&
+      nextProps.groups === this.props.groups
     )
   }
 
@@ -27,7 +30,9 @@ export default class GroupRows extends Component {
       groupHeights,
       onRowClick,
       onRowDoubleClick,
-      clickTolerance
+      clickTolerance,
+      groups,
+      horizontalLineClassNamesForGroup
     } = this.props
     let lines = []
 
@@ -39,6 +44,8 @@ export default class GroupRows extends Component {
           onDoubleClick={evt => onRowDoubleClick(evt, i)}
           key={`horizontal-line-${i}`}
           isEvenRow={i % 2 === 0}
+          group={groups[i]}
+          horizontalLineClassNamesForGroup={horizontalLineClassNamesForGroup}
           style={{
             width: `${canvasWidth}px`,
             height: `${groupHeights[i] - 1}px`