+++ src/lib/utility/calendar.js
@@ -68,7 +68,7 @@ export const minCellWidth = 17
 
 export function getMinUnit(zoom, width, timeSteps) {
   // for supporting weeks, its important to remember that each of these
-  // units has a national progression to the other. i.e. a year is 12 months
+  // units has a natural progression to the other. i.e. a year is 12 months
   // a month is 24 days, a day is 24 hours.
   // with weeks this isnt the case so weeks needs to be handled specially
   let timeDividers = {
@@ -222,7 +222,7 @@ export function collision(a, b, lineHeight, collisionPadding = EPSILON) {
   )
 }
 
-export function stack(items, groupOrders, lineHeight, force) {
+export function stack(items, groupOrders, lineHeight) {
   var i, iMax
   var totalHeight = 0
 
@@ -230,13 +230,6 @@ export function stack(items, groupOrders, lineHeight, force) {
   var groupTops = []
 
   var groupedItems = getGroupedItems(items, groupOrders)
-
-  if (force) {
-    // reset top position of all items
-    for (i = 0, iMax = items.length; i < iMax; i++) {
-      items[i].dimensions.top = null
-    }
-  }
 
   groupedItems.forEach(function(group) {
     // calculate new, non-overlapping positions
@@ -290,7 +283,7 @@ export function stack(items, groupOrders, lineHeight, force) {
   }
 }
 
-export function nostack(items, groupOrders, lineHeight, force) {
+export function nostack(items, groupOrders, lineHeight) {
   var i, iMax
 
   var totalHeight = 0
@@ -299,13 +292,6 @@ export function nostack(items, groupOrders, lineHeight, force) {
   var groupTops = []
 
   var groupedItems = getGroupedItems(items, groupOrders)
-
-  if (force) {
-    // reset top position of all items
-    for (i = 0, iMax = items.length; i < iMax; i++) {
-      items[i].dimensions.top = null
-    }
-  }
 
   groupedItems.forEach(function(group) {
     // calculate new, non-overlapping positions